Replace channel value's data type from Integer to Byte
The Integer data type used for storing DMX channel values has been changed to Byte in order to optimise memory usage. The change impacted several files including the TableStep, StepsViewController, and ScenePacket.java among others. Casting and conversion adjustments were also implemented accordingly.

+++ src/main/java/de/standaloendmx/standalonedmxcontrolpro/gui/bottombar/fader/FaderViewController.java
@@ -74,9 +74,9 @@ public class FaderViewController implements Initializable {
         }
     }
 
-    public void setSliders(Map<Integer, Integer> channelValues) {
+    public void setSliders(Map<Integer, Byte> channelValues) {
 
-        for (Map.Entry<Integer, Integer> entry : channelValues.entrySet()) {
+        for (Map.Entry<Integer, Byte> entry : channelValues.entrySet()) {
             MyFader slider = (MyFader) hBox.getChildren().get(entry.getKey());
             slider.setManaged(true);
             slider.setVisible(true);
